Port fixes

+++ physx/source/foundation/windows/FdWindowsSync.cpp
@@ -67,7 +67,7 @@ void PxSyncImpl::set()
 
 bool PxSyncImpl::wait(uint32_t milliseconds)
 {
-	if(milliseconds == -1)
+	if(milliseconds == static_cast<uint32_t>(-1))
 		milliseconds = INFINITE;
 
 	return WaitForSingleObject(getSync(this), milliseconds) == WAIT_OBJECT_0;
